Renamed SearchArtist components folder to SearchArtists
