+++ src/components/SearchArtists/ArtistSimplifiedCard.module.scss
