+++ src/components/SearchArtists/ArtistsList.jsx
