+++ src/components/SearchArtists/ArtistsList.module.scss
