+++ src/components/SearchArtists/LoadingArtistSmallCard.jsx
