+++ src/components/SearchArtists/LoadingArtistSmallCard.module.scss
