+++ src/components/SearchArtists/LoadingArtistsList.module.scss
