+++ src/components/SearchArtists/index.js
